add tests

+++ src/main/java/project/services/dto/OrderLineDto.java
@@ -3,13 +3,14 @@ package project.services.dto;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Data
 public class OrderLineDto {
 
     @NotNull
-    private BigInteger price;
+    private BigDecimal price;
 
     @NotNull
     private BigInteger weight;
